% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_calls.R
\name{analysis_lasso}
\alias{analysis_lasso}
\title{Run Lasso analysis}
\usage{
analysis_lasso(
  geno_file,
  pheno_file,
  pheno_name,
  out_file,
  bed = TRUE,
  plink_path = TRUE,
  h2 = 0.5
)
}
\arguments{
\item{geno_file}{string specifying path to genotypes file, including file
extension.}

\item{pheno_file}{string specifying path to phenotypes file, including file
extension.}

\item{pheno_name}{column name of phenotype to be used in analysis.}

\item{out_file}{string specifying path and name of the resulting output
file, excluding file extension.}

\item{bed}{logical indicating whether or not the genotypes file is a .bed
file. \code{FALSE} if the file is .ped.}

\item{plink_path}{\code{TRUE} if user has added PLINK to the system
variable "Path". Otherwise a string specifying the path to PLINK.}

\item{h2}{heritability parameter (used for calibration of lasso).}
}
\value{
Does not return anything, but PLINK writes \code{out_file} to disk,
containing results of the performed lasso-regression. The columns in the
printed dataset depends on which phenotype is used for the analysis.
}
\description{
to be written.
NOTE: File extension will be .lasso.
}
