% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phenotypes.R
\name{assign_GWAX_phenotype}
\alias{assign_GWAX_phenotype}
\title{Create phenotype status by proxy}
\usage{
assign_GWAX_phenotype(pheno_file, output_file = pheno_file, sibs)
}
\arguments{
\item{pheno_file}{path to file with information on phenotype status, i.e.
the file "phenotypes.txt" generated by simulation functions.}

\item{output_file}{path of output file, including file extension ".txt".}

\item{sibs}{\emph{optional:} specify number of siblings to use for proxy.
Defaults to using all of an individual's siblings.}
}
\value{
Does not return anything, but writes \code{output_file} to disk,
where the GWAX phenotype column is appended to the end of \code{pheno_file}.
}
\description{
Assigns a new phenotype status to individuals by looking at
family members, for use in Genome Wide Association Studies by proxy (GWAX).
}
\details{
An individual, who is registered as a control, will be assigned case, if any
of its family members are cases. Otherwise the phenotype status is the
same.\cr
\emph{Note:} The `sibs` argument can be used for limiting how many siblings
are taken into consideration when creating the new phenotype This allows for
different analyses with the same simulation, see for instance
\code{vignette("geneference")}.
}
