% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phenotypes.R
\name{assign_ltfh_phenotype}
\alias{assign_ltfh_phenotype}
\title{Calculate posterior mean genetic liabilities}
\usage{
assign_ltfh_phenotype(pheno_file, output_file = pheno_file, alpha, sibs)
}
\arguments{
\item{pheno_file}{path to file with information on liabilities, with same
format as "phenotypes.txt" generated by the simulation functions in
geneference.}

\item{output_file}{path of output file. This is by default the same as
\code{pheno_file}.}

\item{alpha}{significance level used for deciding phenotype status based on
liabilities.}

\item{sibs}{\emph{optional:} number of siblings in the families to be used
for calculating posterior mean genetic liabilities. Defaults to using all
of an individual's siblings.}
}
\value{
Does not return anything, but writes a \code{output_file} to disk,
where configuration, configuration class and posterior mean genetic
liability is appended to the end of \code{pheno_file}.
}
\description{
to be written. Note: sibs argument can be used for using a
dataset with many siblings as a dataset with fewer siblings. Include this
in \code{vignette("geneference")}.
}
