% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{augment_results}
\alias{augment_results}
\title{Augment results for plotting}
\usage{
augment_results(data, alpha)
}
\arguments{
\item{data}{data.table from \code{load_results()}.}

\item{alpha}{significance level used for hypothesis tests.}
}
\value{
A data.table containing the new appended results.
}
\description{
Takes data from \code{load_results()} and appends
columns used for plotting results.
}
\details{
This function requires that this data has been loaded with
\code{load_results()}. \cr
It generates new columns for all columns that represent p-values generated by
\code{analysis_association()}. One of the columns, having "_significant" as
its suffix, denotes whether or not the p-value is significant at the specified
alpha level. The other column, having "_bonferroni" as suffix, denotes whether the p-value is significant
with Bonferroni-corrected alpha level.\cr
Furthermore, the user is prompted to decide if they want to create a column
named "causal", which denotes whether or not a SNP is truly causal for the
phenotype status of an individual. The user needs to have loaded "beta.txt"
with \code{load_results()} in order to create this column.\cr
Lastly, the user is prompted to decide if they want to create a column
named "LTFH_transformed", which is the estimated effect sizes of a
regression using the LT-FH phenotype transformed to the liability scale.\cr
Transforming the effect sizes requires that the user has loaded results from
\code{analysis_association()} with both \code{pheno_name = "line_pheno"} and
\code{pheno_name = "LTFH_pheno"}. These need to be laoded with \code{load_results()}
along with \code{MAFs.txt}.
}
