% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cov.R
\name{calculate_cov}
\alias{calculate_cov}
\title{Calculate covariances between liabilities}
\usage{
calculate_cov(pheno, sibs)
}
\arguments{
\item{pheno}{dataset with phenotypes, as loaded by \code{load_phenotypes()}.}

\item{sibs}{\emph{optional:} number of siblings to be included in covariance
matrix. Defaults to using all of an individual's siblings.}
}
\value{
A covariance matrix using the data from the given dataset.
}
\description{
Compute a sample covariance matrix. This can be used to
validate that covariances of data are as expected.
}
