% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_types.R
\name{calculate_error_types}
\alias{calculate_error_types}
\title{Calculate the number of error types in an analysis}
\usage{
calculate_error_types(dataset, alpha, analysis_name)
}
\arguments{
\item{dataset}{data loaded with \code{load_assoc_results()}.}

\item{alpha}{significance level used for hypothesis tests.}

\item{analysis_name}{name of analysis.}
}
\value{
A tibble with rows... columns...
}
\description{
description to be written.
}
