% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{Compare_beta}
\alias{Compare_beta}
\title{Plot transformed linear GWAS beta values against the LTFH_transformed
values.}
\usage{
Compare_beta(
  dataset,
  beta_x,
  beta_y,
  bonferroni,
  save_plot_path = FALSE,
  plot_filename = "Compare_beta.png"
)
}
\arguments{
\item{dataset}{data imported to R.}

\item{beta_x}{beta-values to be displayed on the x-axis. E.g. the transformed
LTFH beta values.}

\item{beta_y}{beta-values to be displayed on the y-axis. E.g. the linear GWAS
beta values.}

\item{bonferroni}{corrected significant values from LTFH.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
description to be written.
}
