% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_tester.R
\name{cov_tester}
\alias{cov_tester}
\title{cov_tester}
\usage{
cov_tester(file, sib = 0)
}
\arguments{
\item{file}{path to the file that should be read}

\item{sib}{number of siblings. Default is 0}
}
\value{
A covariance matrix using the data from the selected file. All
entries in the matrix are multiplied by 100.
}
\description{
Compute a sample covariance matrix. This can be used to
validate that covariances of data are as expected.
}
