% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmatrix.R
\name{covmatrix}
\alias{covmatrix}
\title{covmatrix}
\usage{
covmatrix(sib = 0, hsq)
}
\arguments{
\item{sib}{is the amount siblings to the entety}

\item{hsq}{heritability parameter that is squard}
}
\value{
This function returns the covariance matrix for an entety with sib - siblings.
}
\description{
This function create a covariance matrix for a entety regarding the heritability.
The first row is full liability for a entety, next is genetic liability.
Row 3 and 4 is the entetys parents liability.
}
\examples{
covmatrix(2, 0.5)

}
