% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{create_map}
\usage{
create_map(n, path)
}
\arguments{
\item{n}{is the amount of sequenced enteties.}

\item{path}{The path where the file is going to be stored.}
}
\value{
This function creates a map file
}
\description{
This function creates a map file that is necessary for PLINK.

Note change the first entry "data_map[, 1]" to have cromosones in the analysis.
}
