% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_for_testsim}
\alias{data_for_testsim}
\title{simulated data}
\format{
A list of 4 elements.
}
\source{
Randomly generated data
}
\usage{
data_for_testsim
}
\description{
A simulated dataset using the testsim function with seed 42.
The code for this is: testsim(10, 15, 5, 0.5, 0.05, TRUE)
}
\keyword{datasets}
