% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_dist_simulaiton.R
\name{family_dist_simulaiton}
\alias{family_dist_simulaiton}
\title{Simulation with varying family sizes}
\usage{
family_dist_simulaiton(n, m, q, hsq, k, dist, path = "")
}
\arguments{
\item{n}{number of genotypes (individuals). Given as a vector that match the
dist vector (see discribtion below).}

\item{m}{number of SNPS per genotype.}

\item{q}{number of causal SNPs, i.e. SNPs that effect chances of having
the phenotype.}

\item{hsq}{squared heritability parameter.}

\item{k}{prevalence of phenotype.}

\item{dist}{the distribution of siblings. Given as a vector with the same length
as the "n"-vector.}

\item{path}{directory where the files will be stored. If nothing is
specified, \code{family_simulation} writes its files in the current
working directory.}
}
\value{
Does not return any value, but prints the following five files to
the \code{path} parameter specified in the function call:
\itemize{
\item Three text files:
\itemize{
\item beta.txt - a file of \code{m} rows with one column. The i'th row is
the true effect of the i'th SNP.
\item MAFs.txt - a file of \code{m} rows with one column. The i'th row is
the true Minor Allelle Frequence of the i'th SNP.
\item phenotypes.txt - a file of \code{n} rows, number of columns depend on
number of siblings. Note that some individuals has fewer siblings as
others. Thus we fill the empty value-slots with "-9". The file
contains the phenotype and liability of
each individual as well as information on the liabilities and phenotype
status of their parents and siblings.
}
\item genotypes.map - a file created such that PLINK will work with the genotype
data.
\item genotypes.ped - the simulated genotypes in a PLINK-readable format.
}
}
\description{
Simulate genotypes for individuals with family history, where
each individual has an arbitrary number of siblings. Parents' genotypes are
simulated and used for simulating the genotypes of individuals and their
siblings.
\code{family_dist_simulaiton} makes use of parallel computation in order to
decrease the running time.
}
\details{
Note that dist denotes the amount of siblings matched with the amount of
individuals n. E.g. n = c(100, 200, 300, 400) and dist = c(0, 1, 2, 3) would mean
that we want a total of 100 + 200 + 300 + 400 = 1000 individuals. Here 100
individuals have 0 siblings, 200 has 1 sibling and so on...
}
\section{Warning}{

Simulating large datasets lead to very large files (eg. in the order of 100k
genotypes with 100k SNPs takes up approximately 40GB of space).
Please ensure that you have sufficient disk space at the disk where
\code{path} resides prior to running the simulation.
See FUNCTION-NAME DOCUMENTATION REFERENCE(this is a placeholder until
function is implemented) in order to convert the generated .ped to the
smaller .bed format.
}

