% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_simulation.R
\name{family_simulation}
\alias{family_simulation}
\title{family_simulation}
\usage{
family_simulation(n, m, q, hsq, k, path = "", sib = 0)
}
\arguments{
\item{n}{is the amount of enteties (humans)}

\item{m}{is the amount of SNPS per entety}

\item{q}{is the amount of causal SNPs}

\item{hsq}{heritability parameter that is squard}

\item{k}{is the prevalence of trait parameter}

\item{path}{is the file path, where the files will be stored. If nothing is added, the function will place the function
at the place of the working directory. Use getwd() to see your working directory}

\item{sib}{is number of siblings per individual.}
}
\value{
This function returns five files: Three txt files: Beta, MAFs and phenotypes, a genotypes MAP file and a genotypes PED file.
}
\description{
This function simulate data for individual enteties with specefic family history. both the liability and
pheno types of both siblings and parents are known to the subjects.
}
