% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{GWAS_beta_VS_LTFH_transfored_beta}
\alias{GWAS_beta_VS_LTFH_transfored_beta}
\title{Plot transformed linear GWAS beta values against the LTFH_transformed
values.}
\usage{
GWAS_beta_VS_LTFH_transfored_beta(
  dataset,
  transformed_beta,
  linear_beta,
  bonferroni,
  save_plot_path = FALSE,
  plot_filename = "LTFH.png"
)
}
\arguments{
\item{dataset}{data imported to R.}

\item{transformed_beta}{the transformed beta values from LTFH.}

\item{linear_beta}{the beta values from standard GWAS.}

\item{bonferroni}{corrected significant values from LTFH.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
description to be written.
}
