% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_assoc_results}
\alias{load_assoc_results}
\title{Import and merge simulation data with results of analysis}
\usage{
load_assoc_results(plink_file, true_effects_file, MAFs_file, alpha)
}
\arguments{
\item{plink_file}{path to the assoc or qassoc file generated by
\code{analysis_association()}. Include file extension.}

\item{true_effects_file}{path of file with true effects used in simulation,
including file extension.
The simulation functions save this file as "beta.txt".}

\item{MAFs_file}{path of file with true effects used in simulation,
including file extension.
The simulation functions save this file as "MAFs.txt".}

\item{alpha}{significance level used for single SNP tests.}
}
\value{
A tibble where true values of effects and MAFs for each SNP has been
merged with the results from the analysis performed by PLINK. In addition,
the data is augmented with the logical columns
\itemize{
\item \code{causal}: \code{TRUE} if the SNP is causal for the phenotype,
\item \code{significant}: \code{TRUE} if the effect of the SNP was significant,
\item \code{bonferroni}: \code{TRUE} if the effect of the SNP was significant
with Bonferroni-correction.
These columns are used in in REFERENCE FUNCTIONS(error types and plots)
}
}
\description{
Placeholder description: Merge results from PLINK assoc with
values used for simulation.
Note that paths can be specified as relative to working directory.
}
