% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_phenotypes}
\alias{load_phenotypes}
\title{Import phenotypes from simulation}
\usage{
load_phenotypes(pheno_file)
}
\arguments{
\item{pheno_file}{path of file with phenotypes, including file extension.
Simulation functions save this file as "phenotypes.txt".}
}
\value{
Returns a tibble with the data from \code{pheno_file}.
}
\description{
Import phenotypes from simulation
}
