% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_results}
\alias{load_results}
\title{Import, merge and augment results}
\usage{
load_results(path)
}
\arguments{
\item{path}{path to the folder containing the data, including "/" or "\\\\"
at the end. The function can not take relative paths, i.e. the user always
needs to specify the full path to the directory.}
}
\value{
A data.table containing the different results from the files.
}
\description{
Merge results from analysis performed in PLINK with the true values used by
the simulation.
}
\details{
The function only takes a directory as input and then prompts the user to
decide which files in the directory should be loaded. Therefore, simulation
data and results from \code{analysis_association()} should reside in the
same directory in order to be merged.\cr
The names of the columns are created by adding the name of the file as
suffix. E.g., the P-values from an analysis named \code{GWAS.assoc} will be
in the column named "P_GWAS".\cr
Note that this function is not supposed to load phenotype information, but
rather loads data relating directly to the SNPs. See
\code{load_phenotypes()} for reading phenotype related data.\cr
}
