% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results.R
\name{merge_results}
\alias{merge_results}
\title{Title}
\usage{
merge_results(folder_path, assoc, true_effects = "beta.txt", m)
}
\arguments{
\item{assoc}{}

\item{true_effects}{}

\item{m}{}
}
\value{
something
}
\description{
Title
}
