% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{mutator}
\alias{mutator}
\title{This function takes the data from the reader-function and does append
some calculated columns.}
\usage{
mutator(data, alpha)
}
\arguments{
\item{data}{data.table from the reader function.}

\item{alpha}{significance level (often set as 0.05).}
}
\value{
A data.table containing the new appended results.
}
\description{
This function assumes that all columns starting with "P_" are
columns with p-values. The function creates to new column per the "P_"
columns. One column called "..._significnat", which indicates what SNPs are
viewed as significant according to the analysis. The function also creates a
"..._bonferroni" that does the same, exept this column uses the bonferroni
correctsion. The function also ask the user if they want to make a
transformed LTFH-column. If the user have not done the LTFH-analysis, then
they most say no (0) to this action. If yes (1) is pressed, then the function
will ask the user which columns the function shall use for the analysis.
}
