% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_sibs.R
\name{n_sibs}
\alias{n_sibs}
\title{Find number of siblings in largest family}
\usage{
n_sibs(pheno)
}
\arguments{
\item{pheno}{tibble loaded with load_pheno().}
}
\value{
Number of siblings in the largest family.
}
\description{
Find number of siblings in largest family
}
