% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_calls.R
\name{p_2_b}
\alias{p_2_b}
\title{Convert .ped files to .bed}
\usage{
p_2_b(ped_file, bed_file = ped_file, del = TRUE, plink_path = TRUE)
}
\arguments{
\item{ped_file}{path of the .ped file to be converted, excluding file
extension.}

\item{bed_file}{path of the .bed file to be written, excluding file
extension.}

\item{del}{if \code{TRUE}, deletes the \code{ped_file} after conversion.}

\item{plink_path}{\code{TRUE} if user has added PLINK to the system
variable "PATH". Otherwise a string specifying the path to the folder
containing plink.exe.}
}
\value{
Does not return anything, but writes the converted \code{ped_file}
to disk at the path \code{bed_file}.
}
\description{
to be written. Essentially:
calls plink to convert for storing with smaller file size.
NOTE: .ped file can be viewed, .bed cannot be viewed
}
