% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_validation.R
\name{path_validation}
\alias{path_validation}
\title{Check if certain files exists at a path}
\usage{
path_validation(path)
}
\arguments{
\item{path}{for validation}
}
\value{
A choice of a valid path, new folder, over-writing the files or
stopping the simulation.
}
\description{
This function check if certain files exists at the path that is
given to any of the simulation functions. If these files exsists, the user
has different options: 1. Stopping the simulation, 2. Over-writing the files
(deleting them and creating new ones), 3. Choose a new path for the files
and 4. Make a new folder at the given path, for the new files. If the user
does not give valid input the function will re-ask the user.
}
