% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_pmgl_vs_true}
\alias{plot_pmgl_vs_true}
\title{Plot true genetic liabilities against posterior mean genetic
liabilities}
\usage{
plot_pmgl_vs_true(
  dataset,
  LTFH_pheno = LTFH_pheno,
  child_lg = child_lg,
  line_color = "black",
  line_type = "dashed",
  label_size = 3,
  save_plot_path = FALSE,
  plot_filename = "posterior_liabilities.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_assoc_results()}.}

\item{LTFH_pheno}{the phenotypes calculted for LT-FH.}

\item{child_lg}{the genetic liability of the individual.}

\item{line_color}{color of identity line.}

\item{line_type}{type of identity line.}

\item{label_size}{the size of the boxes.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
write description.
\code{vignette("ggplot2-specs", package = "ggplot2")} details the usage of
aesthetic parameters in ggplot2.
}
