% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_pval_hist}
\alias{plot_pval_hist}
\title{Plot a histogram of p-values}
\usage{
plot_pval_hist(
  dataset,
  bins = 20,
  line_color = "black",
  fill_color = "cornflowerblue",
  mean_color = "red",
  save_plot_path = FALSE,
  plot_filename = "pvalue_histogram.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_assoc_results()}.}

\item{bins}{number of bins used.}

\item{line_color}{outline color of bins.}

\item{fill_color}{fill color of bins.}

\item{mean_color}{color of the line showing the mean value.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
write something
\code{vignette("ggplot2-specs", package = "ggplot2")} details the usage of
aesthetic parameters in ggplot2.
}
