% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{reader}
\alias{reader}
\title{Import and merge analysis results from plink and txt-files as MAF}
\usage{
reader(path, vars = c("P", "BETA", "CHISQ", "SE"))
}
\arguments{
\item{path}{full path to the folder containing the analysis-files.}

\item{vars}{variables of interest that is to be imported if they exist in
the current files.}
}
\value{
A data.table containing the different results from the files.
}
\description{
Placeholder description: Merge results from PLINK with
values used for simulation. The names of the columns are created from the
names of the different files. Note that this function is nt supposed to read
the different pheno-files. The function does not let the user select a file
if the file contain the string "pheno". If there is a pheno-file at the
path not containing "pheno", then please do not select it. This function
assumes that the .txt files only contain one coulmn. Using the simulations
this would be the beta and MAFs files.
}
