% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnorm.R
\name{rmnorm}
\alias{rmnorm}
\title{rmnorm}
\usage{
rmnorm(n, S)
}
\arguments{
\item{n}{is the amount of random vectors that is to be simulated.}

\item{S}{is a valid covariance matrix.}
}
\value{
This function returns a matrix with n rows and i columns according to the amount of columns in S.
}
\description{
Simulate random values form a multi-variate normal distribution.
}
