% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fixed_family.R
\name{sim_fixed_family}
\alias{sim_fixed_family}
\title{Simulation with fixed family size}
\usage{
sim_fixed_family(n, m, q, hsq, k, sib = 0, path = "")
}
\arguments{
\item{n}{number of genotypes (individuals).}

\item{m}{number of SNPS per genotype.}

\item{q}{number of causal SNPs, i.e. SNPs that effect chances of having
the phenotype.}

\item{hsq}{squared heritability parameter.}

\item{k}{prevalence of phenotype.}

\item{sib}{number of siblings per individual.}

\item{path}{directory where the files will be stored. If nothing is
specified, \code{sim_fixed_family} writes its files in the current
working directory.}
}
\value{
Does not return any value, but prints the following five files to
the \code{path} parameter specified in the function call:
\itemize{
\item Three text files:
\itemize{
\item beta.txt - a file of \code{m} rows with one column. The i'th row is
the true effect of the i'th SNP.
\item MAFs.txt - a file of \code{m} rows with one column. The i'th row is
the true Minor Allelle Frequency of the i'th SNP.
\item phenotypes.txt - a file of \code{n} rows, number of columns depend on
number of siblings. The file contains the phenotype status and liability
of each individual as well as information on the liabilities and
phenotype status of their parents and siblings.
}
\item genotypes.map - a file created such that PLINK will work with the genotype
data.
\item genotypes.ped - the simulated genotypes in a PLINK-readable format.
Note: The function only saves genotype data for the target individual.
}
}
\description{
Simulate genetic data, including genotypes, phenotype status
and liabilities, for individuals and their family, where each individual
has a fixed number of siblings specified by the \code{sib} parameter.
}
\details{
Parents' genotypes are simulated and used for creating the genotypes of
the individuals and their siblings. For the methodology behind the
simulation, see \code{vignette("liability-distribution")}.\cr
\code{sim_fixed_family} makes use of parallel computation in order to
decrease the running time. As one CPU core is left unused, the user
should be able to do other work while the simulation is running.
}
\section{Warning}{

Simulating large datasets takes time and generates large files. For details
on time complexity and required disk space, see
\code{vignette("sim-benchmarks")}.\cr
The largest file generated is \code{genotypes.ped}. See \code{convert_geno_file()} to convert it
to another file format, thereby reducing its size significantly.
}

