% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_random_family.R
\name{sim_random_family}
\alias{sim_random_family}
\title{Simulation with random family sizes}
\usage{
sim_random_family(n, m, q, hsq, k, sib_fert, dist = 0, path = "")
}
\arguments{
\item{n}{number of genotypes (individuals).}

\item{m}{number of SNPS per genotype.}

\item{q}{number of causal SNPs, i.e. SNPs that effect chances of having
the phenotype.}

\item{hsq}{squared heritability parameter.}

\item{k}{prevalence of phenotype.}

\item{sib_fert}{either the distribution vector or a fertility rate. See
details section.}

\item{dist}{if \code{sib_fert} is a distribution vector, then \code{dist} is used to
specify how many siblings the probabilities in \code{sib_fert} correspond to.}

\item{path}{directory where the files will be stored. If nothing is
specified, \code{sim_random_family} writes its files in the current
working directory.}
}
\value{
A list where first entry is the number of individuals, i.e. the \code{n}
parameter supplied to \code{sim_varied_family()}, and the second entry is the
number of siblings those individuals have, i.e. the  \code{dist} parameter
supplied to \code{sim_varied_family()}.\cr
This is returned once the simulation is done printing the following five
files to the \code{path} parameter specified in the function call:
\itemize{
\item Three text files:
\itemize{
\item beta.txt - a file of \code{m} rows with one column. The i'th row is
the true effect of the i'th SNP.
\item MAFs.txt - a file of \code{m} rows with one column. The i'th row is
the true Minor Allelle Frequency of the i'th SNP.
\item phenotypes.txt - a file of \code{n} rows, number of columns depend on
number of siblings. The file contains the phenotype status and liability
of each individual as well as information on the liabilities and
phenotype status of their parents and siblings.
}
\item genotypes.map - a file created such that PLINK will work with the genotype
data.
\item genotypes.ped - the simulated genotypes in a PLINK-readable format.
Note: The function only saves genotype data for the target individual.
}
}
\description{
Simulate genetic data, including genotypes,
phenotype status and liabilities, for individuals with family history, where
each individual is given a random number of siblings.
}
\details{
Parents' genotypes are simulated and used for creating the genotypes of
the individuals and their siblings. For the methodology behind the
simulation, see \code{vignette("liability-distribution")}.\cr
The number of siblings generated by either a multinomial distribution or
a Poisson distribution. The function then calls \code{sim_varied_family()}
with the appropriate \code{n} and \code{dist} parameters.\cr
The choice of distribution depends on the input given to \code{sib_fert}. If a
a single number is given, the function uses the Poisson distribution to
randomly select how many siblings the different individuals have. In this
case, \code{dist} should not be specified.\cr
If a vector is given, then the function uses \code{sib_fert} as probabilities
of a multinomial distribution to select how many individuals have the number
of siblings specified in each entry of \code{dist}.\cr
E.g., \code{sib_fert = c(1/6, 1/6, 2/3)} and \code{dist = c(0, 2, 3)} means that the
probability of having 0 siblings is 1/6, the probability of having 2
siblings is 1/6 and the probability of having 3 siblings is 2/3.\cr
Since individuals have a different number of siblings, some entries in
\code{phenotypes.txt} will be missing, denoted by -9.\cr
E.g., an individual with 1 sibling in a dataset where the maximum number of
siblings is 3, would have -9 in all columns relating to sibling 2 and
sibling 3.\cr
\code{sim_random_family} makes use of parallel computation in order to
decrease the running time. As one CPU core is left unused, the user
should be able to do other work while the simulation is running.
}
\section{Warning}{

Simulating large datasets takes time and generates large files. For details
on time complexity and required disk space, see
\code{vignette("sim-benchmarks")}.\cr
The largest file generated is \code{genotypes.ped}. See \code{convert_geno_file()} to convert it
to another file format, thereby reducing its size significantly.
}

