% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test.R
\name{sim_test}
\alias{sim_test}
\title{Simulation directly in the R session without family history}
\usage{
sim_test(n, m, q, hsq, k, to_ped = T)
}
\arguments{
\item{n}{number of genotypes (individuals).}

\item{m}{number of SNPS per genotype.}

\item{q}{number of causal SNPs, i.e. SNPs that effect chances of having
the phenotype.}

\item{hsq}{squared heritability parameter.}

\item{k}{prevalence of phenotype.}

\item{to_ped}{if `TRUE`, write the genotype output in PLINK's .ped format.
Else, write the number of minor alleles for each genotype at each SNP.}
}
\value{
Returns a list with 5 entries, containing a matrix of genotypes, a
vector specifying effect sizes of SNPs, a vector with liability of
individual i, a vector specifying phenotype  status of individual i, and the
transformed phenotype status used for linear regression in PLINK.
}
\description{
Simulate genetic data in R, including genotypes,
phenotype status and liabilities, for individuals.
}
\details{
A quick way of simulating a small amount of data to be stored
in the R-session. Note that this function only simulates data for
individuals without family history, i.e. they are not simulated as
offspring. Hence, the simulation cannot be used in a study of family
history.
}
\section{Warning}{

Since this function saves the generated data directly in R (that is, on RAM),
the function should not be used for large simulations, but rather to get an
idea of the general framework.\cr
The function does not run in parallel.
}

\examples{
sim_test(10, 15, 5, 0.5, 0.05, TRUE)
}
