% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Testsim.R
\name{testsim}
\alias{testsim}
\title{testsim}
\usage{
testsim(n, m, q, hsq, k, to_ped = T)
}
\arguments{
\item{n}{is the amount of enteties (humans)}

\item{m}{is the amount of SNPS per entety}

\item{q}{is the amount of causal SNPs}

\item{hsq}{heritability parameter that is squard}

\item{k}{is the prevalence of trait parameter}

\item{to_ped}{is a T/F parameter indicatating if the output should be in a plink-friendly format or in binary}
}
\value{
This function returns a list containing a matrix of genotypes, causal SNPs, the enteties liabilities and the phenotypes
}
\description{
A quick way of simulating a small amount of data in the R-session. Note that this function
only simulate data for individuals. So the simulation should not be used in a study with siblings/parents.
}
\examples{
testsim(10, 15, 5, 0.5, 0.05, TRUE)


}
