% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/To_ped.R
\name{to_ped}
\alias{to_ped}
\title{to_ped}
\usage{
to_ped(df, part)
}
\arguments{
\item{df}{input of binary matrix}

\item{part}{is the amount of parts, any given function splits the data in}
}
\value{
This function retruns a n x m matrix
}
\description{
turns a binary matrix into a PLINK-friendly format
}
