% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_calls.R
\name{analysis_association}
\alias{analysis_association}
\title{Run association analysis}
\usage{
analysis_association(
  geno_file,
  pheno_file,
  pheno_name,
  out_file,
  bed = TRUE,
  plink_path = TRUE
)
}
\arguments{
\item{geno_file}{string specifying path to genotypes file, including file
extension.}

\item{pheno_file}{string specifying path to phenotypes file, including file
extension.}

\item{pheno_name}{column name of phenotype to be used in analysis.}

\item{out_file}{string specifying path and name of the resulting output
file, excluding file extension.}

\item{bed}{logical indicating whether or not the genotypes file is a .bed
file. \code{FALSE} if the file is .ped.}

\item{plink_path}{\code{TRUE} if user has added PLINK to the system
variable "PATH". Otherwise a string specifying the path to the folder
containing plink.exe.}
}
\value{
Does not return anything, but PLINK writes \code{out_file} to disk,
containing results of the performed analysis. The file format of
\code{out_file} and the columns contained in it depends on which phenotype
is used for the analysis.
}
\description{
Calls PLINK to run an association analysis on specified data.
}
\details{
PLINK uses the \code{pheno_name} column to determine which kind of analysis
to run on the data.\cr
Use "pheno" for a 1 degree of freedom chi-square test with the case-control
representation of the phenotype.\cr
Use "line_pheno" for a linear regression using the case-control
representation of the phenotype.\cr
Use "LTFH_pheno" for a linear regression using the LTFH phenotype. Requires
that \code{pheno_file} was created/edited by \code{assign_ltfh_phenotype()}.
Use "GWAX_pheno" for a linear regression using the LTFH phenotype. Requires
that \code{pheno_file} was created/edited by \code{assign_GWAX_phenotype()}.
}
