% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_calls.R
\name{analysis_association}
\alias{analysis_association}
\title{Run association analysis}
\usage{
analysis_association(
  geno_file,
  pheno_file,
  pheno_name,
  out_file,
  bed = TRUE,
  plink_path = TRUE
)
}
\arguments{
\item{geno_file}{string specifying path to genotypes file, including file
extension.}

\item{pheno_file}{string specifying path to phenotypes file, including file
extension.}

\item{pheno_name}{column name of phenotype to be used in analysis.}

\item{out_file}{string specifying path and name of the resulting output
file, excluding file extension.}

\item{bed}{logical indicating whether or not the genotypes file is a .bed
file. \code{FALSE} if the file is .ped.}

\item{plink_path}{\code{TRUE} if user has added PLINK to the system
variable "PATH". Otherwise a string specifying the path to the folder
containing plink.exe.}
}
\value{
Does not return anything, but PLINK writes \code{out_file} to disk,
containing results of the performed analysis. The columns in the printed
dataset depends on which phenotype is used for the analysis.
}
\description{
to be written. calls PLINK.
(include specifying pheno_names from simulation:
GWAS use pheno, for LTFH use LTFH_pheno, GWAX use GWAX_pheno)
NOTE: creates a temporary .bed file if \code{bed} is \code{FALSE}. This
takes some additional time.
Also: File extension will be .assoc or .qassoc depending on the used model.
}
