% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phenotypes.R
\name{assign_GWAX_phenotype}
\alias{assign_GWAX_phenotype}
\title{Create case-control status by proxy}
\usage{
assign_GWAX_phenotype(pheno_file, output_file = pheno_file, sibs = 0)
}
\arguments{
\item{pheno_file}{path to file with information on phenotype status, with
same format as "phenotypes.txt" generated by the simulation functions in
geneference.}

\item{output_file}{path of output file. This can be the same as
\Code{pheno_file}.}

\item{sibs}{number of siblings in the families. If family size varies,
specify the number of siblings in the largest family.}
}
\value{
Does not return anything, but writes \code{output_file} to disk,
where the GWAX phenotype column is appended to the end of \code{pheno_file}.
}
\description{
to be written
}
