% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phenotypes.R
\name{assign_GWAX_phenotype}
\alias{assign_GWAX_phenotype}
\title{Create case-control status by proxy}
\usage{
assign_GWAX_phenotype(pheno_file, output_file = pheno_file, sibs)
}
\arguments{
\item{pheno_file}{path to file with information on phenotype status, with
same format as "phenotypes.txt" generated by the simulation functions in
geneference.}

\item{output_file}{path of output file. This is by default the same as
\code{pheno_file}.}

\item{sibs}{\emph{optional:} specify number of siblings to use for proxy.
Defaults to using all of an individual's siblings.}
}
\value{
Does not return anything, but writes \code{output_file} to disk,
where the GWAX phenotype column is appended to the end of \code{pheno_file}.
}
\description{
to be written. Note: sibs argument can be used for using a
dataset with many siblings as a dataset with fewer siblings. Include this
in \code{vignette("geneference")}.
}
