% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_phenotypes.R
\name{assign_ltfh_phenotype}
\alias{assign_ltfh_phenotype}
\title{Calculate posterior mean genetic liabilities}
\usage{
assign_ltfh_phenotype(pheno_file, sibs, output_file = pheno_file, alpha)
}
\arguments{
\item{pheno_file}{path to file with information on liabilities, with same
format as "phenotypes.txt" generated by the simulation functions in
geneference.}

\item{sibs}{number of siblings in the families. If family size varies,
specify the number of siblings in the largest family.}

\item{output_file}{path of output file. This is by default the same as
\code{pheno_file}.}

\item{alpha}{significance level used for deciding phenotype status based on
liabilities.}
}
\value{
Does not return anything, but writes a \code{output_file} to disk,
where configuration, configuration class and posterior mean genetic
liability is appended to the end of \code{pheno_file}.
}
\description{
to be written
}
