% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cov.R
\name{calculate_cov}
\alias{calculate_cov}
\title{Calculate covariance matrix of liabilities}
\usage{
calculate_cov(pheno, sibs)
}
\arguments{
\item{pheno}{dataset with phenotypes, as loaded by \code{load_phenotypes()}.}

\item{sibs}{\emph{optional:} number of siblings to be included in covariance
matrix. Defaults to using all of an individual's siblings.}
}
\value{
A covariance matrix using the data from \code{pheno}.
}
\description{
Compute a sample covariance matrix from \code{"phenotypes.txt"}
as generated by simulation functions.
}
\details{
The outputted matrix follows the order described in
\code{vignette("liability-distribution")}.The output can be used to validate
that covariances between the liabilities of family members are as
expected.\cr
The \code{sibs} parameter can be used to adjust how many siblings to
calculate the covariance between.
}
