% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{compare_beta}
\alias{compare_beta}
\title{Plot transformed linear GWAS beta values against the LTFH_transformed
values.}
\usage{
compare_beta(
  dataset,
  beta_x,
  beta_y,
  bonferroni,
  line_color = "green",
  line_type = "dashed",
  save_plot_path = FALSE,
  plot_filename = "Compare_beta.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_results()}.}

\item{beta_x}{column containing beta-values to be displayed on the x-axis.
E.g. the transformed LTFH beta values. State in the format dataset$beta_x.}

\item{beta_y}{column containing beta-values to be displayed on the y-axis.
E.g. the linear GWAS beta values. State in the format dataset$beta_y.}

\item{bonferroni}{column containing the corrected significant values from LTFH.
State in the format dataset$bonferroni.}

\item{line_color}{the color of the identity line.}

\item{line_type}{the type of the identity line.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.
Must be either .pdf, .png or .jpeg}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
Creates a scatterplot of transformed linear GWAS beta values against the LTFH_transformed values.
}
\details{
In the plot an identity line is included, which helps visualise how well
the estimations compare to each other. \cr
The plot requires the user to run an analysis with
\code{analysis_association()}, load the results and merge with
\code{beta.txt} using \code{load_results()} and augment results with
\code{augment_results()} choosing to create the LTFH transformed column. \cr
\code{vignette("ggplot2-specs", package = "ggplot2")} details the usage of
aesthetic parameters in ggplot2.
}
