% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_tester.R
\name{cov_tester}
\alias{cov_tester}
\title{cov_tester}
\usage{
cov_tester(read, sib = 0)
}
\arguments{
\item{read}{the path to the file that should be read}

\item{sib}{is the amount of siblings}
}
\value{
A covariance matrix using th data from the selected file
}
\description{
This function simulate data for individual enteties (the data for the parents)
is unknown.
}
