% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmatrix.R
\name{covmatrix}
\alias{covmatrix}
\title{Create covariance matrix for liabilities}
\usage{
covmatrix(hsq, sib = 0)
}
\arguments{
\item{hsq}{heritability parameter.}

\item{sib}{number of siblings.}
}
\value{
A covariance matrix for the liabilities of a family with \code{sib}
number of siblings.
}
\description{
Design the theoretical covariance matrix from the multivariate
normal distribution used to model the liabilities.
}
\details{
The covariance matrix is explained in detail in
\code{vignette("liability-distribution")}.\cr
}
\examples{
covmatrix(0.5, 2)

}
