% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmatrix.R
\name{covmatrix}
\alias{covmatrix}
\title{covmatrix}
\usage{
covmatrix(sib = 0, hsq)
}
\arguments{
\item{sib}{number of siblings}

\item{hsq}{squared heritability parameter}
}
\value{
A covariance matrix for an individual with sib number of siblings,
such that the number of rows and columns are equal to 4 + number of
siblings.
}
\description{
Design the theoretical covariance matrix for the multivariate
normal used to model the distribution of liabilities.
The first entry in the vector of liabilities is the genetic liability, the
next is the full liability and the two following are the liabilities of the
parents. Liabilities of siblings constitute the remaining entries in the
vector.
}
\examples{
covmatrix(2, 0.5)

}
