% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_estimates_vs_true}
\alias{plot_estimates_vs_true}
\title{Plot true effects against estimated effects of SNPs}
\usage{
plot_estimates_vs_true(
  dataset,
  save_plot_path = FALSE,
  plot_filename = "beta_comparison.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_assoc_results()}.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
description to be written.
}
