% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot of p-values}
\usage{
plot_manhattan(
  dataset,
  save_plot_path = FALSE,
  plot_filename = "manhattan_plot.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_assoc_results()}.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
description to be written. Plots a Manhattan plot with only the
SNPs that are significant at a 5%-significance level.
Write a few sentences about why the plot is different from real world plots
(linkage disequilibrium)
}
