% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_manhattan}
\alias{plot_manhattan}
\title{Manhattan plot of p-values}
\usage{
plot_manhattan(
  dataset,
  P,
  SNP = SNP,
  causal = causal,
  line_color = "cornflowerblue",
  line_type = "dashed",
  save_plot_path = FALSE,
  plot_filename = "manhattan_plot.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_results()}.}

\item{P}{column containing p-values to be plotted. State in the format
dataset$P_value.}

\item{SNP}{column containing the SNP values. If loaded with \code{load_results()}
can be left default otherwise state in the format dataset$SNP.}

\item{causal}{column containing the causal values. If created created with
\code{augment_results()} can be left default, otherwise state in the format
dataset$causal.}

\item{line_color}{color of the lines.}

\item{line_type}{the type of the lines.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.
Must be either .pdf, .png or .jpeg}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
Plots a Manhattan plot with only the
SNPs that have a p-value of at most 0.1.
}
\details{
All points above the upper dashed line show SNPs that are significant at
0.05 significance level with Bonferroni-correction. Points above the lower
dashed line show SNPs that are significant at a 0.05 significance level.\cr
The color of the point show whether or not they are truly causal for the
phenotype.\cr
The user must run some analysis with \code{analysis_association()} 
and load the data with \code{load_results()}. Furthermore, \code{augment_results()}
must be run where the user should create a 'causal' column
from the true effect sizes.
}
