% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_pmgl_vs_true}
\alias{plot_pmgl_vs_true}
\title{Plot true genetic liabilities against posterior mean genetic
liabilities}
\usage{
plot_pmgl_vs_true(
  dataset,
  line_color = "black",
  line_type = "dashed",
  label_size = 3,
  save_plot_path = FALSE,
  plot_filename = "posterior_liabilities.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_phenotypes()}.}

\item{line_color}{color of identity line.}

\item{line_type}{type of identity line.}

\item{label_size}{the size of the boxes.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.
Must be either .pdf, .png or .jpeg}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
Creates a scatterplot of genetic liability against posterior mean genetic liability.
}
\details{
Visualises how well the posterior mean genetic liabilities fit the genetic liabilities.
An identity line is included to help see if the genetic liabilities are balanced above
and below for each posterior mean genetic liability. Furthermore, the color and labels
show each configuration class, which is detailed in
\code{vignette("liability-distribution")}. \cr
This function requires the user to run \code{assign_ltfh_phenotype()}, and then
\code{association_analysis(pheno_name = "LTFH_pheno")}. Data is loaded with
\code{load_phenotypes()}. \cr
\code{vignette("ggplot2-specs", package = "ggplot2")} details the usage of
aesthetic parameters in ggplot2.
}
