% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisations.R
\name{plot_pmgl_vs_true}
\alias{plot_pmgl_vs_true}
\title{Plot true genetic liabilities against posterior mean genetic
liabilities}
\usage{
plot_pmgl_vs_true(
  dataset,
  line_color = "black",
  save_plot_path = FALSE,
  plot_filename = "posterior_liabilities.png"
)
}
\arguments{
\item{dataset}{data imported to R by \code{load_assoc_results()}.}

\item{line_color}{color of identity line.}

\item{save_plot_path}{if \code{FALSE}, return the function returns a ggplot
object. Else a path of the directory to save the plot to.}

\item{plot_filename}{name of the file to be saved, including file extension.}
}
\value{
Either returns a \code{ggplot} object or saves the plot to
\code{save_plot_path} and returns NULL.
}
\description{
write description.
\code{vignette("ggplot2-specs", package = "ggplot2")} details the usage of
aesthetic parameters in ggplot2.
}
