% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmnorm.R
\name{rmnorm}
\alias{rmnorm}
\title{Multivariate normal distributions}
\usage{
rmnorm(n, S)
}
\arguments{
\item{n}{number of random vectors that is to be simulated.}

\item{S}{a valid covariance matrix.}
}
\value{
Returns a matrix with n rows and i columns, where i is determined by
the size of the specified covariance matrix.
}
\description{
Simulate random values from a multivariate normal distribution.
}
\examples{
rmnorm(n = 10, S = covmatrix(sib = 0, hsq = 0.5))
}
