% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_no_family.R
\name{sim_no_family}
\alias{sim_no_family}
\title{sim_no_family}
\usage{
sim_no_family(n, m, q, hsq, k, path)
}
\arguments{
\item{n}{is the amount of enteties (humans)}

\item{m}{is the amount of SNPS per entety}

\item{q}{is the amount of causal SNPs}

\item{hsq}{heritability parameter that is squard}

\item{k}{is the prevalence of trait parameter}

\item{path}{is the file path, where the files will be stored}
}
\value{
This function returns five files: Three txt files: Beta, MAFs and phenotypes, a genotypes MAP file and a genotypes PED file.
}
\description{
This function simulate data for individual enteties (the data for the parents)
is unknown.
}
