% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_random_family.R
\name{sim_random_family}
\alias{sim_random_family}
\title{Simulation with random family sizes}
\usage{
sim_random_family(n, m, q, hsq, k, sib_fert, dist = 0, path = "")
}
\arguments{
\item{n}{number of genotypes (individuals).}

\item{m}{number of SNPS per genotype.}

\item{q}{number of causal SNPs, i.e. SNPs that effect chances of having
the phenotype.}

\item{hsq}{squared heritability parameter.}

\item{k}{prevalence of phenotype.}

\item{sib_fert}{either the distribution-vector of siblings-fertility rate.
See meaning below.}

\item{dist}{if sib_fert is a distribution-vector then dist can be used to
specify how many siblings the different individuals get.}

\item{path}{directory where the files will be stored. If nothing is
specified, \code{family_simulation} writes its files in the current
working directory.}
}
\value{
a list where first entry is the distribution of the simulated individuals
and the second entry is the number of siblings (dist) for these individuals.
the function also call family_dist_simulaiton that prints the following five files to
the \code{path} parameter specified in the function call:
\itemize{
\item Three text files:
\itemize{
\item beta.txt - a file of \code{m} rows with one column. The i'th row is
the true effect of the i'th SNP.
\item MAFs.txt - a file of \code{m} rows with one column. The i'th row is
the true Minor Allelle Frequence of the i'th SNP.
\item phenotypes.txt - a file of \code{n} rows, number of columns depend on
number of siblings. Note that some individuals has fewer siblings as
others. Thus we fill the empty value-slots with "-9". The file
contains the phenotype and liability of
each individual as well as information on the liabilities and phenotype
status of their parents and siblings.
}
\item genotypes.map - a file created such that PLINK will work with the genotype
data.
\item genotypes.ped - the simulated genotypes in a PLINK-readable format.
}
}
\description{
Simulate genotypes for individuals with family history, where
each individual is given a random number of siblings. Parents' genotypes are
simulated and used for simulating the genotypes of individuals and their
siblings. This function call the family_dist_simulaiton with the randomized
input.
\code{family_dist_simulaiton} makes use of parallel computation in order to
decrease the running time.
}
\details{
Note that there are two types of input to sib_fert:
distribution-vector or siblings-fertility. If a distribution-vector is used
a dist vector must also be specified. E.g. sib_fert = c(1/6, 1/6, 2/3)
and dist = c(0, 1, 2). In this case the function is asked to randomly choose
1/6 individuals to have 0 siblings and 2/3 to have 2 siblings. The other case
a value is given to sib_fert. E.g. sib_fert = 2. Here the function use the
poisson distribution to randomly select how many siblings the different
individuals has. Note that in this case, no dist has to be specified, since
the function overwrites it with the range of the simulated siblings.
}
\section{Warning}{

Simulating large datasets lead to very large files (eg. in the order of 100k
genotypes with 100k SNPs takes up approximately 40GB of space).
Please ensure that you have sufficient disk space at the disk where
\code{path} resides prior to running the simulation.
See FUNCTION-NAME DOCUMENTATION REFERENCE(this is a placeholder until
function is implemented) in order to convert the generated .ped to the
smaller .bed format.
}

