% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_test.R
\name{sim_test}
\alias{sim_test}
\title{Simulate genotypes in the R session without family history}
\usage{
sim_test(n, m, q, hsq, k, to_ped = T)
}
\arguments{
\item{n}{number of genotypes (individuals).}

\item{m}{number of SNPS per genotype.}

\item{q}{number of causal SNPs, i.e. SNPs that effect chances of having
the phenotype.}

\item{hsq}{squared heritability parameter.}

\item{k}{prevalence of phenotype.}

\item{to_ped}{TRUE/FALSE indicating if the output should be in a
PLINK-friendly format or not.}
}
\value{
Returns a list of 4 entries, containing a matrix of genotypes, a
vector specifying indexes of causal SNPs, a vector the liability of
individual i, and a vector specifying whether individual i has the
phenotype.
}
\description{
A quick way of simulating a small amount of data to be stored
in the R-session. Note that this function only simulates data for
individuals without family history, i.e. they are not simulated as
offspring. Hence, the simulation cannot be used in a study of family
history.
}
\examples{
testsim(10, 15, 5, 0.5, 0.05, TRUE)

}
