% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/To_ped.R
\name{to_ped}
\alias{to_ped}
\title{to_ped}
\usage{
to_ped(df, part)
}
\arguments{
\item{df}{genotypes in our simulated format}

\item{part}{amount of parts any given function splits the data in}
}
\value{
Returns a n x (6 + 2 * m) matrix so it follows PLINK's .ped format.
}
\description{
Converts our format to PLINK's .bed-format.
}
